turned off gsl error handling

+++ src/SampAlpha.h
@@ -21,7 +21,7 @@ void alpha_samp(const int nobs, const int K, const int P, const int L,
 
     for(l=0; l<L; l++){
         Ml = BF.Ml[l];
-        double * tempmm = (double *)calloc(Ml*Ml, sizeof(double));
+        float * tempmm = (float *)calloc(Ml*Ml, sizeof(float));
         gsl_vector * alpha_mean = gsl_vector_alloc(Ml);
         gsl_matrix * alpha_samp = gsl_matrix_alloc(1, Ml);
         gsl_matrix * alpha_prec = gsl_matrix_alloc(Ml, Ml);
